requirements.txt
GitPython==3.1.44
langchain==0.3.21
langchain_ollama==0.2.3